backend/requirements.txt
django
djangorestframework
django-cors-headers
gunicorn
whitenoise
python-dotenv
psycopg2-binary
dj-database-url
